Fix MyPy type errors in experiment scripts
- Convert np.mean() results to float() in leniency_experiment.py and collusive_leniency_experiment.py
- Resolves dict type mismatch: numpy.floating[Any] vs float
- Ensures GitHub workflow passes all type checks

+++ scripts/leniency_experiment.py
@@ -432,8 +432,8 @@ def run_leniency_experiment(
             "avg_consumer_surplus": float(
                 np.mean([ep["consumer_surplus"] for ep in episodes])
             ),
-            "avg_producer_surplus": np.mean(
-                [ep["producer_surplus"] for ep in episodes]
+            "avg_producer_surplus": float(
+                np.mean([ep["producer_surplus"] for ep in episodes])
             ),
         }
